+++ crates/ruff_linter/resources/test/fixtures/fastapi/FAST001.py
