+++ crates/ruff_linter/resources/test/fixtures/fastapi/FAST002.py
