+++ crates/ruff_linter/src/rules/mod.rs
@@ -1,6 +1,7 @@
 #![allow(clippy::useless_format)]
 pub mod airflow;
 pub mod eradicate;
+pub mod fastapi;
 pub mod flake8_2020;
 pub mod flake8_annotations;
 pub mod flake8_async;
