+++ crates/ruff_linter/src/rules/ruff/mod.rs
@@ -1,6 +1,5 @@
 //! Ruff-specific rules.
 
-mod fastapi;
 pub(crate) mod rules;
 pub(crate) mod typing;
 
@@ -57,8 +56,6 @@ mod tests {
     #[test_case(Rule::UnusedAsync, Path::new("RUF029.py"))]
     #[test_case(Rule::AssertWithPrintMessage, Path::new("RUF030.py"))]
     #[test_case(Rule::RedirectedNOQA, Path::new("RUF101.py"))]
-    #[test_case(Rule::FastApiRedundantResponseModel, Path::new("RUF102.py"))]
-    #[test_case(Rule::FastApiNotAnnotatedDependency, Path::new("RUF103.py"))]
     fn rules(rule_code: Rule, path: &Path) -> Result<()> {
         let snapshot = format!("{}_{}", rule_code.noqa_code(), path.to_string_lossy());
         let diagnostics = test_path(
