+++ crates/ruff_linter/src/rules/ruff/rules/mod.rs
@@ -5,8 +5,6 @@ pub(crate) use asyncio_dangling_task::*;
 pub(crate) use collection_literal_concatenation::*;
 pub(crate) use default_factory_kwarg::*;
 pub(crate) use explicit_f_string_type_conversion::*;
-pub(crate) use fastapi_not_annotated_dependency::*;
-pub(crate) use fastapi_redundant_response_model::*;
 pub(crate) use function_call_in_dataclass_default::*;
 pub(crate) use implicit_optional::*;
 pub(crate) use invalid_formatter_suppression_comment::*;
@@ -40,8 +38,6 @@ mod collection_literal_concatenation;
 mod confusables;
 mod default_factory_kwarg;
 mod explicit_f_string_type_conversion;
-mod fastapi_not_annotated_dependency;
-mod fastapi_redundant_response_model;
 mod function_call_in_dataclass_default;
 mod helpers;
 mod implicit_optional;
