+++ ruff.schema.json
@@ -3061,6 +3061,11 @@
         "FA10",
         "FA100",
         "FA102",
+        "FAST",
+        "FAST0",
+        "FAST00",
+        "FAST001",
+        "FAST002",
         "FBT",
         "FBT0",
         "FBT00",
@@ -3681,8 +3686,6 @@
         "RUF10",
         "RUF100",
         "RUF101",
-        "RUF102",
-        "RUF103",
         "RUF2",
         "RUF20",
         "RUF200",
